requirements.txt
Django==5.2.8
djangorestframework==3.16.1
python-dateutil==2.9.0.post0
